tests/web_platform_tests/graph/identity-create.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>DID Identity — Credential Management Extension — WPT</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
'use strict';

// ==========================================================================
// Spec 02 — Decentralised Identity Integration for the Web Platform
// https://github.com/HexaField/w3c-living-web-proposals/blob/main/drafts/02_decentralised-identity-web-platform.md
// ==========================================================================

// §3.1 DIDCredential interface

promise_test(async t => {
  assert_true('credentials' in navigator, 'navigator.credentials MUST exist');
  assert_true(typeof navigator.credentials.create === 'function');
  assert_true(typeof navigator.credentials.get === 'function');
}, '§3.2 navigator.credentials MUST support create() and get()');

promise_test(async t => {
  // Note: In production, create() requires a user gesture.
  // In WPT test mode, user gesture requirements may be relaxed.
  const credential = await navigator.credentials.create({
    did: { displayName: 'Test Identity' }
  });

  assert_true(credential instanceof DIDCredential, 'MUST return DIDCredential');
  assert_equals(credential.type, 'did', 'type MUST be "did"');
  assert_true(credential.did.startsWith('did:key:'), 'MUST be did:key');
  assert_equals(credential.algorithm, 'Ed25519', 'MUST use Ed25519');
  assert_equals(credential.displayName, 'Test Identity');
  assert_true(credential.createdAt.length > 0, 'MUST have createdAt');
  assert_false(credential.isLocked, 'MUST not be locked initially');
}, '§3.2 credentials.create({ did }) MUST return DIDCredential with correct properties');

promise_test(async t => {
  const c1 = await navigator.credentials.create({
    did: { displayName: 'ID1' }
  });
  const c2 = await navigator.credentials.create({
    did: { displayName: 'ID2' }
  });
  assert_not_equals(c1.did, c2.did, 'DIDs MUST be unique');
}, '§3.2 Each created DIDCredential MUST have a unique DID');

// §3.3 Retrieving a DIDCredential

promise_test(async t => {
  await navigator.credentials.create({ did: { displayName: 'Retrievable' } });
  const retrieved = await navigator.credentials.get({ did: {} });
  assert_true(retrieved instanceof DIDCredential);
  assert_true(retrieved.did.startsWith('did:key:'));
}, '§3.3 credentials.get({ did }) MUST retrieve an existing DIDCredential');

// §5.1 sign()

promise_test(async t => {
  const credential = await navigator.credentials.create({
    did: { displayName: 'Signer' }
  });
  const data = { message: 'Hello', number: 42 };
  const signed = await credential.sign(data);

  assert_equals(signed.author, credential.did, 'author MUST be signer DID');
  assert_true(signed.timestamp.length > 0, 'MUST have timestamp');
  assert_equals(signed.proof.key, credential.did, 'proof.key MUST be DID');
  assert_true(signed.proof.signature.length > 0, 'MUST have signature');
}, '§5.1 sign() MUST return SignedContent with author, timestamp, proof');

// §5.2 verify()

promise_test(async t => {
  const credential = await navigator.credentials.create({
    did: { displayName: 'Verifier' }
  });
  const signed = await credential.sign({ test: true });
  assert_true(await credential.verify(signed), 'MUST verify valid signature');
}, '§5.2 verify() MUST return true for valid signature');

promise_test(async t => {
  const credential = await navigator.credentials.create({
    did: { displayName: 'TamperTest' }
  });
  const signed = await credential.sign({ test: true });
  const tampered = {
    author: signed.author,
    timestamp: signed.timestamp,
    data: { test: false },
    proof: signed.proof
  };
  assert_false(await credential.verify(tampered), 'MUST reject tampered data');
}, '§5.2 verify() MUST return false for tampered content');

// §4.3.2 Lock and Unlock

promise_test(async t => {
  const credential = await navigator.credentials.create({
    did: { displayName: 'Lockable' }
  });
  await credential.lock();
  assert_true(credential.isLocked, 'MUST be locked after lock()');

  await promise_rejects_dom(t, 'InvalidStateError',
    credential.sign({ data: 'test' }),
    'sign() MUST reject while locked');

  await credential.unlock();
  assert_false(credential.isLocked, 'MUST be unlocked after unlock()');
}, '§4.3.2 lock() MUST prevent signing, unlock() MUST restore it');

// §3.4 Supported Algorithms

promise_test(async t => {
  const credential = await navigator.credentials.create({
    did: { displayName: 'Ed25519 Test', algorithm: 'Ed25519' }
  });
  assert_equals(credential.algorithm, 'Ed25519', 'MUST support Ed25519');
}, '§3.4 User agent MUST support Ed25519 algorithm');
</script>
